+++ tools/server/inference.py
@@ -1,22 +1,4 @@
-import queue
-from http import HTTPStatus
-from pathlib import Path
 
-import numpy as np
-import torch
-from kui.asgi import HTTPException
-from loguru import logger
-
-from fish_speech.text.chn_text_norm.text import Text as ChnNormedText
-from fish_speech.utils import autocast_exclude_mps, set_seed
-from tools.file import AUDIO_EXTENSIONS, audio_to_bytes, list_files, read_ref_text
-from tools.inference_engine.utils import decode_vq_tokens, wav_chunk_header
-from tools.llama.generate import (
-    GenerateRequest,
-    GenerateResponse,
-    WrappedGenerateResponse,
-)
-from tools.schema import ServeTTSRequest
 
 
 @torch.inference_mode()
